server/requirements.txt
Flask==2.3.2
waitress
redis==5.0.0
python-dotenv==1.0.0